Comercio listar, editar, eliminar, crear, buscar

+++ XeonComerce/WebUI/ClientApp/src/app/comercios/comercios.component.ts
@@ -1,3 +1,6 @@
+import { ConfirmDialogComponent } from './../_components/confirm-dialog/confirm-dialog.component';
+import { filter } from 'rxjs/operators';
+import { Direccion } from './../_models/direccion';
 import { Ubicacion } from './../_models/ubicacion';
 import { Comercio } from '../_models/comercio';
 import { Component, OnInit, Inject, ViewChild } from '@angular/core';
@@ -7,6 +10,7 @@ import { UbicacionService } from '../_services/ubicacion.service';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import {MatTableDataSource } from '@angular/material/table';
 import {MatSort} from '@angular/material/sort';
+import { DireccionService } from '../_services/direccion.service';
 
 @Component({
   selector: 'app-comercios',
@@ -22,8 +26,9 @@ export class ComerciosComponent implements OnInit {
 	provincias: Ubicacion[];
 	cantones: Ubicacion[];
 	distritos: Ubicacion[];
+	direccion: Direccion;
 	
-	constructor(public dialog: MatDialog, private comercioService: ComercioService, private mensajeService: MensajeService, private ubicacionService: UbicacionService) { }
+	constructor(public dialog: MatDialog, private comercioService: ComercioService, private direccionService: DireccionService, private mensajeService: MensajeService, private ubicacionService: UbicacionService) { }
   
 
 	@ViewChild(MatSort, {static: true}) sort: MatSort;
@@ -77,8 +82,57 @@ export class ComerciosComponent implements OnInit {
 		console.log(`Resultado: ${result}`); 
 		if (result) {
 			//Revisar si están los datos de dir y crearla retornar el id y ya.
-		  this.comercioService.create(result)
-			.subscribe(() => this.getComercios());
+			console.log("Primero se crea una dirección");
+			
+
+			this.direccion = {
+				id: -1,
+				provincia: result.provincia,
+				canton: result.canton,
+				distrito: result.distrito,
+				sennas: result.sennas,
+				latitud: result.lat,
+				longitud: result.long
+			}
+
+			let direccionFinal: Direccion
+			this.direccionService.create(this.direccion).subscribe(() => {
+				this.direccionService.get()
+				.subscribe(dirs => {
+					dirs = dirs.filter((a) => {
+						return (a.provincia == this.direccion.provincia && a.canton == this.direccion.canton && a.distrito == this.direccion.distrito && a.latitud == this.direccion.latitud && a.longitud == this.direccion.longitud && a.sennas == this.direccion.sennas);
+					  });
+					direccionFinal = dirs[0];
+					if(direccionFinal){
+						console.log("Direccion a registrar es: ", this.direccion);
+						let comercioFinal : Comercio;
+						comercioFinal = {
+							"cedJuridica": result.cedJuridica,
+							"nombreComercial": result.nombreComercial,
+							"correoElectronico": result.correoElectronico,
+							"telefono": result.telefono,
+							"direccion": direccionFinal.id,
+							"idUsuario": result.idUsuario
+						}
+					this.comercioService.create(comercioFinal)
+					.subscribe(() => {
+
+						
+			this.provincias = [];
+			this.cantones = [];
+			this.distritos = [];
+			this.direccion = undefined;
+			this.getProvincias();
+
+						this.getComercios()
+					});
+					}else{
+						console.log("Algo ocurrio al buscar el id de la direccion.");
+					}
+
+				});
+
+			});
 		}
   
 	  });
@@ -136,8 +190,18 @@ export class ComerciosComponent implements OnInit {
   
   
 	delete(comercio: Comercio): void {
-	  this.comercioService.delete(comercio)
-		.subscribe(() => this.getComercios());
+
+		const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+			maxWidth: "500px",
+			data: {
+				title: "Está seguro?",
+				message: "Usted está apunto de eliminar un comercio. "}
+		  });
+		
+		  dialogRef.afterClosed().subscribe(dialogResult => {
+			  if(dialogResult) this.comercioService.delete(comercio)
+			  .subscribe(() => this.getComercios());
+		 });
 	}
   
 }
@@ -151,7 +215,7 @@ export class ComerciosComponent implements OnInit {
   
 	constructor(
 	  public dialogRef: MatDialogRef<DialogComercio>,
-	  @Inject(MAT_DIALOG_DATA) public data: Comercio, 
+	  @Inject(MAT_DIALOG_DATA) public data: any, 
 	  private ubicacionService: UbicacionService) { }
   
 	onNoClick(): void {
@@ -159,7 +223,7 @@ export class ComerciosComponent implements OnInit {
 	}
 
 	
-	getCantonesE(event: Event): void {
+	getCantonesE(event: any): void {
 		console.log("Recibido");
 		let provincia = event.value;
 		console.log(event.value)
@@ -168,7 +232,7 @@ export class ComerciosComponent implements OnInit {
 	  }
 
 	
-	getDistritosE(event: Event): void {
+	getDistritosE(event: any): void {
 		console.log("Recibido");
 		let canton = event.value;
 		console.log(event.value)
